hararecity-councile/requirements.txt
langchain-core>=0.1.40
langchain>=0.1.0
langchain-community>=0.0.13
langchain-groq>=0.1.0
streamlit==1.31.0
python-dotenv==1.0.0 